Primitive VCL tokenizer
Thanks to a lack of escaping in VCL strings and a composition of tokens
that never goes beyond two characters, it's a good-enough approximation.
There are some two-character tokens like '!~' or '<=' that are seen as
two individual tokens (respectively '!' and '~' or '<' and '=') but this
is not a problem until the preprocessor needs to care about them (and it
is very unlikely to happen). Even with a three-character token it would
still be possible to use this construct.

The important thing is not to confuse the tokenizer and start looking
for tokens inside strings or comments...