Collect the test suite coverage
There is currently no way to collect code coverage of Rust code using
the upstream tooling (like Go does for example with `go test`). Even if
it was possible, this would likely be done via `cargo` and unit tests
(probably something similar to `go test -cover`).

But with a test suite relying on `varnishtest` that would most likely
not play well together. Interestingly `kcov` is capable of line coverage
with no compile-time instrumentation, and it can consolidate results
from several runs (in this case several VTCs).

The wrapper script is here to keep the VTCs simple and not leak coverage
details in test cases. So not only is compile-time instrumentation
avoided thanks to `kcov` but test cases can be run both:

- in the test suite (with or without `kcov`)
- against a local installation of `vclpp`

This script breaks the directory isolation provided by `varnishtest` but
makes it much more easier to merge results from multiple VTCs without
having to worry about parallel executions.