Distinction between NeedsMore and MayNeedMore
PVCL code may end on an unbounded token (in the test case, an integer)
and complain about incomplete VCL. NeedsMore is now for tokens that
rely on a separator to finish (eg. /* C comments */ or "strings") or
for unbounded tokens when the current character isn't allowed to be the
last one.