wip: Expand a token to zero or more
Instead of writing code directly, synthetic tokens are created when the
alternative syntax is expanded. Some tokens like comments are lost, some
expand to more than one. Once the resulting tokens are connected, only
then they are written to the output.

This commit breaks more than half of the test suite, however it is a
necessary step to reach the goal of complete segregation between writing
to the output and preprocessing.