Kill unnecessary copies of tokens
Oh boy! This one sure had its share of learning curve.

So the general design of the preprocessor is now a matryoshka of lazy
iterators feeding from each others in a single thread. So why copy
tokens around between the tokenizer and the passes? Well, the only
cases are bad and synthetic tokens, because we don't want to mutate
the originals if any. Otherwise let all the actors share tokens with
a reference counter: RcToken.

Because of re-scoping happening in a couple places, this patch is
probably best viewed with:

    git show --ignore-all-space --word-diff --word-diff-regex=. $SHA1

Token is now immutable, and will only create reference-counted copies
when synthetic tokens are derived.