Introduce aliases for imported VMODs
This time, learning a bit about string manipulation, or rather not
learning much. There's probably a better way to reconstruct original
VMOD uses when hitting an alias. Rust's built-in HashMap is rather
nice although most of its capabilities are most likely overkill for
vclpp's needs since only a handful of VMODs are expected to be loaded
at once. Sill... a convenient turn-key interace, so why not using it?