Fold the bracket check into the flow
This gets rid of an awkward pile up of passes and still gets to check
both the input PVCL and the resulting VCL output for balance, after each
pass.

Moving the responsibility also gives the opportunity to consolidate
"incomplete VCL" errors once and for all, hopefully. It also greatly
reduces the hoops when dealing with failures in the passes.

When running with kcov, the flow also checks that once it sees a None,
it will see a None afterwards. Returning Some after None is not strictly
bad iterator behavior, and it can be worked around using an adapter, but
thou shalt not need a fuse. That too lives in one place now.

Back to a 100%-ish coverage.